reorganize survey conditional test setup to share with survey patch tests
