+++ test/fixtures/conditional-session/conditional/index.js
@@ -1,0 +1,15 @@
+'use strict';
+
+const setup = require('./setup');
+const requiredOverrides = require('./required-overrides');
+const passAnswer = require('./pass-answer');
+const errorAnswer = require('./error-answer');
+const userSurveys = require('./user-surveys');
+
+module.exports = {
+    setup,
+    requiredOverrides,
+    passAnswer,
+    errorAnswer,
+    userSurveys,
+};
