+++ test/survey_conditional.model.spec.js
@@ -18,6 +18,8 @@ const History = require('./util/history');
 const SharedSpec = require('./util/shared-spec');
 const choiceSetCommon = require('./util/choice-set-common');
 const surveyCommon = require('./util/survey-common');
+const conditionalSession = require('./fixtures/conditional-session/conditional');
+const choiceSets = require('./fixtures/example/choice-set');
 
 const expect = chai.expect;
 
@@ -25,22 +27,29 @@ describe('conditional survey unit', function surveyConditionalUnit() {
     const hxUser = new History();
     const hxSurvey = new SurveyHistory();
     const hxChoiceSet = new History();
+    const counts = [8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8, 8];
 
     const answerer = new Answerer();
     const questionGenerator = new QuestionGenerator();
-    const surveyGenerator = new CSG({ questionGenerator, answerer, hxSurvey });
+    const surveyGenerator = new CSG({
+        questionGenerator,
+        answerer,
+        hxSurvey,
+        setup: conditionalSession.setup,
+        requiredOverrides: conditionalSession.requiredOverrides,
+        counts,
+    });
 
     const generator = new Generator({ surveyGenerator, questionGenerator, answerer });
     const shared = new SharedSpec(generator);
 
-    const numOfCases = surveyGenerator.numOfCases();
+    const numOfCases = counts.length;
 
     const tests = new surveyCommon.SpecTests(generator, hxSurvey);
     const choiceSetTests = new choiceSetCommon.SpecTests(generator, hxChoiceSet);
 
     before(shared.setUpFn());
 
-    const choiceSets = CSG.getChoiceSets();
     choiceSets.forEach((choiceSet, index) => {
         it(`create choice set ${index}`, choiceSetTests.createChoiceSetFn(choiceSet));
         it(`get choice set ${index}`, choiceSetTests.getChoiceSetFn(index));
@@ -91,7 +100,7 @@ describe('conditional survey unit', function surveyConditionalUnit() {
         it(`create user ${index}`, shared.createUserFn(hxUser));
     });
 
-    CSG.conditionalErrorSetup().forEach((errorSetup) => {
+    conditionalSession.errorAnswer.forEach((errorSetup) => {
         it(`error: survey ${errorSetup.surveyIndex} validation ${errorSetup.caseIndex}`, () => {
             const { surveyIndex, error } = errorSetup;
             const survey = hxSurvey.server(surveyIndex);
@@ -107,7 +116,7 @@ describe('conditional survey unit', function surveyConditionalUnit() {
         });
     });
 
-    CSG.conditionalPassSetup().forEach((passSetup) => {
+    conditionalSession.passAnswer.forEach((passSetup) => {
         let answers;
 
         it(`create survey ${passSetup.surveyIndex} answers ${passSetup.caseIndex}`, () => {
@@ -152,7 +161,7 @@ describe('conditional survey unit', function surveyConditionalUnit() {
     };
 
     const statusMap = {};
-    CSG.conditionalUserSurveysSetup().forEach((userSurveySetup, stepIndex) => {
+    conditionalSession.userSurveys.forEach((userSurveySetup, stepIndex) => {
         const { skipAnswering, surveyIndex, missingSurveys, status } = userSurveySetup;
         if (!skipAnswering) {
             let answers;
